remove the `quoted` rule
this wasn't very useful anyway because tree-sitter
only chose to use that rule (and not just make the
greedy regex after it in the `value` sequence
take everything) when the quoted string had a
comment token in it.

while it is a shame to not have a highlight for
the quoted strings - it's better not to have it
rather than have it only some of the time (as
described above).